dev > main: updated unit tests
