+++ tests/tests_core/tests_poly.py
@@ -199,69 +199,3 @@ def test_get_integral_coefficients_rote(
         coeff_, coeff, eps=1e-6, message='Derivative should return original coefficients.'
     )
     return
-
-
-def test_get_critical_points(
-    test: TestCase,
-    debug: Callable[..., None],
-    module: Callable[[str], str],
-):
-    # p(t) = t^2 - 2t + 1
-    results = get_critical_points([1, -2, 1])
-    test.assertEquals(len(results), 1)
-    t0, y0, v, kind = results[0]
-    test.assertEquals(t0, 1.0)
-    test.assertEquals(v, 1.0)
-    test.assertEquals(kind, EnumCriticalPoints.LOCAL_MINIMUM)
-
-    # p(t) = -(t^2 - 2t + 1)
-    results = get_critical_points([-1, 2, -1])
-    test.assertEquals(len(results), 1)
-    t0, y0, v, kind = results[0]
-    test.assertEquals(t0, 1)
-    test.assertEquals(v, 1)
-    test.assertEquals(kind, EnumCriticalPoints.LOCAL_MAXIMUM)
-
-    # p(t) = t^3
-    results = get_critical_points([0, 0, 0, 1])
-    test.assertEquals(len(results), 1)
-    t0, y0, v, kind = results[0]
-    test.assertEquals(t0, 0.0)
-    test.assertEquals(v, 2)
-    test.assertEquals(kind, EnumCriticalPoints.INFLECTION)
-
-    # Force
-    #    p'(t) = t^2 - 3t + 2 = (t - 2)(t - 1)
-    # => p''(t) = 2t - 3
-    # so
-    #    p''(1) = -1 --> (1, p(1)) loc. max
-    #    p''(2) = +1 --> (2, p(2)) loc. min
-    coeff = get_integral_coefficients([2, -3, 1])
-    results = get_critical_points(coeff)
-    test.assertEquals(len(results), 2)
-
-    t0, y0, v, kind = results[0]
-    test.assertEquals(t0, 1.0)
-    test.assertEquals(v, 1)
-    test.assertEquals(kind, EnumCriticalPoints.LOCAL_MAXIMUM)
-
-    t0, y0, v, kind = results[1]
-    test.assertEquals(t0, 2.0)
-    test.assertEquals(v, 1)
-    test.assertEquals(kind, EnumCriticalPoints.LOCAL_MINIMUM)
-
-    # Force
-    #    p'(t) = t^2 - 2t + 1 = (t - 1)^2
-    # => p''(t) = 2(t - 1)
-    # => p'''(t) = 2
-    # so
-    #    p''(1) = 0
-    #    p'''(1) > 0 --> (1, p(1)) inflection
-    coeff = get_integral_coefficients([1, -2, 1])
-    results = get_critical_points(coeff)
-    test.assertEquals(len(results), 1)
-    t0, y0, v, kind = results[0]
-    test.assertEquals(t0, 1.0)
-    test.assertEquals(v, 2)
-    test.assertEquals(kind, EnumCriticalPoints.INFLECTION)
-    return
